PAYARA-4176 Documentation and cleanup in CorbaXAITest
